Refactor helper code into separate file.
Also shorten the examples.
